wallet-service/pyproject.toml
[project]
name = "wallet-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Matias Vagliviello",email = "vaglimatias@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)"
]


[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=2.0.0,<3.0.0"]
package-mode = false

[tool.poetry]
name = "wallet-service"
version = "0.1.0"
description = ""
authors = ["Matias"]
package-mode = false
